pyproject.toml
[build-system]
requires = ["setuptools>=75", "setuptools-scm"]
build-backend = "setuptools.build_meta"

[project]
name = "brain"
version = "0.1.0"
description = "Local second-brain: document ingestion, knowledge graph, and RAG chat"
requires-python = ">=3.10"
dependencies = [
    "litellm>=1.55",
    "graphiti-core[anthropic]>=0.8",
    "qdrant-client>=1.12",
    "python-docx>=1.1",
    "python-pptx>=1.0",
    "pydantic>=2.0",
    "pydantic-settings>=2.6",
    "python-dotenv>=1.0",
    "pyyaml>=6.0",
    "typer[all]>=0.12",
    "streamlit>=1.40",
    "rich>=13.0",
    "watchdog>=4.0",
    "voyageai>=0.3",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "pytest-cov>=5.0",
    "ruff>=0.8",
]

[project.scripts]
brain = "brain.cli.app:app"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py310"
